split up fuselage submodels
