+++ 1682/gas_male/submodels/fuselage.py
@@ -1,26 +1,8 @@
 " fuselage.py "
 import numpy as np
 from gpkit import Variable, Model
-
-class FuelTank(Model):
-    """
-    Returns the weight of the fuel tank.  Assumes a cylinder shape with some
-    fineness ratio
-    """
-    def __init__(self, Wfueltot, **kwargs):
-
-        W = Variable("W", "lbf", "fuel tank weight")
-        f = Variable("f", 0.03, "-", "fraction fuel tank weight to fuel weight")
-        mfac = Variable("m_{fac}", 1.1, "-", "fuel volume margin factor")
-        rhofuel = Variable("\\rho_{fuel}", 6.01, "lbf/gallon",
-                           "density of 100LL")
-        Vol = Variable("\\mathcal{V}", "ft^3", "fuel tank volume")
-
-        constraints = [W >= f*Wfueltot,
-                       Vol/mfac >= Wfueltot/rhofuel,
-                      ]
-
-        Model.__init__(self, None, constraints, **kwargs)
+from fuel_tank import FuelTank
+from fuselage_skin import FuselageSkin
 
 class Fuselage(Model):
     "The thing that carries the fuel, engine, and payload"
@@ -59,41 +41,6 @@ class FuselageLoading(Model):
         skinloading = fuselage.skin.loading(fuselage.skin, Wcent)
 
         Model.__init__(self, None, skinloading)
-
-class FuselageSkin(Model):
-    "fuselage skin model"
-    def __init__(self, S, d, l):
-
-        W = Variable("W", "lbf", "fuselage skin weight")
-        g = Variable("g", 9.81, "m/s^2", "Gravitational acceleration")
-        rhokevlar = Variable("\\rho_{kevlar}", 1.3629, "g/cm**3",
-                             "kevlar density")
-        t = Variable("t", "in", "skin thickness")
-        tmin = Variable("t_{min}", 0.03, "in", "minimum skin thickness")
-        I = Variable("I", "m**4", "wing skin moment of inertia")
-
-        self.loading = FuselageSkinL
-
-        constraints = [W >= S*rhokevlar*t*g,
-                       t >= tmin,
-                       I <= np.pi*(d/2)**3*t,
-                       l == l]
-
-        Model.__init__(self, None, constraints)
-
-class FuselageSkinL(Model):
-    "fuselage skin loading"
-    def __init__(self, static, Wcent):
-
-        Mh = Variable("M_h", "N*m", "horizontal axis center fuselage moment")
-        Nmax = Variable("N_{max}", 5, "-", "max loading")
-        sigmakevlar = Variable("\\sigma_{Kevlar}", 190, "MPa",
-                               "stress strength of Kevlar")
-
-        constraints = [Mh >= Nmax*Wcent/4*static["l"],
-                       sigmakevlar >= Mh*static["d"]/2/static["I"]]
-
-        Model.__init__(self, None, constraints)
 
 class FuselageAero(Model):
     "fuselage drag model"
